requirements.txt
streamlit==1.38.0
streamlit-drawable-canvas==0.9.3
opencv-python-headless==4.10.0.84
numpy
pandas
Pillow
matplotlib
